refactor: find all open course
