+++ src/repository/course-open/course-open.repository.ts
@@ -1,12 +1,21 @@
 import { availableCourseItem, InsertAvailableCourseItem } from "@db/schema";
 import { DeleteCourseOpenDto } from "@module/course-open/dto/delete-course-open.dto";
+import { HttpService } from "@nestjs/axios";
 import { Inject, Injectable } from "@nestjs/common";
+import { ConfigService } from "@nestjs/config";
+import { resolveTerm } from "@repository/course-registration/helper/resolveTerm";
 import { Drizzle } from "@type/drizzle.type";
+import { TERM } from "@util/constants";
 import { and, eq } from "drizzle-orm";
+import { firstValueFrom } from "rxjs";
 
 @Injectable()
 export class CourseOpenRepository {
-  constructor(@Inject("DRIZZLE") private drizzle: Drizzle) {}
+  constructor(
+    @Inject("DRIZZLE") private drizzle: Drizzle,
+    private readonly httpService: HttpService,
+    private configService: ConfigService,
+  ) {}
 
   async createCourseOpen(item: InsertAvailableCourseItem) {
     // check if courseId exists
@@ -29,12 +38,6 @@ export class CourseOpenRepository {
     return courseOpen[0];
   }
 
-  async findAllOneTerm(termYearID: number) {
-    return await this.drizzle.query.availableCourseItem.findMany({
-      where: eq(availableCourseItem.availableCourseId, termYearID),
-    });
-  }
-
   async delete(deleteCourseOpenDto: DeleteCourseOpenDto) {
     return this.drizzle
       .delete(availableCourseItem)
@@ -48,5 +51,28 @@ export class CourseOpenRepository {
         ),
       )
       .returning();
+  }
+
+  async findAllOneTerm(term: TERM, year: number) {
+    const termResolved = resolveTerm(term);
+    const url = this.configService.get<string>("service");
+    const { data } = await firstValueFrom(
+      this.httpService.get(
+        `${url}/open_course?term=${termResolved}&year=${year}`,
+      ),
+    );
+    if (!data) {
+      this.httpService.patch(`${url}/open_course`, {
+        term: termResolved,
+        year,
+        majors: [],
+      });
+      return {
+        term: termResolved,
+        year,
+        majors: [],
+      };
+    }
+    return data;
   }
 }
